refactor(inventory): Phase 3 IInventoryStore to Domain/Abstractions/
