+++ inventory-microservice/Inventory.Service/Infrastructure/Data/EntityFramework/EntityFrameworkExtensions.cs
@@ -1,3 +1,4 @@
+using Inventory.Service.Domain.Abstractions;
 using Microsoft.EntityFrameworkCore;
 
 namespace Inventory.Service.Infrastructure.Data.EntityFramework;
