+++ inventory-microservice/Inventory.Service/IntegrationEvents/EventHandlers/ProductCreatedEventHandler.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus.Abstractions;
 using Inventory.Service.Contracts.Integration;
-using Inventory.Service.Infrastructure.Data;
+using Inventory.Service.Domain.Abstractions;
 
 namespace Inventory.Service.IntegrationEvents.EventHandlers;
